moved embeddings to own file
