+++ src/modsim2/similarity/mmd.py
@@ -1,31 +1,9 @@
 import numpy as np
 from sklearn import metrics
 
+from modsim2.similarity.embeddings import EMBEDDING_FN_DICT
 
-def array_to_matrix(array: np.ndarray) -> np.ndarray:
-    """
-    Reshapes arbitrarily sized np.ndarrays to a matrix with the
-    same number of observations
-
-    Args:
-        array (np.ndarray): An np.ndarray with observations on
-                            the first index
-
-    Returns:
-        np.ndarray: An np.ndarray with the same number of observations
-                    reshaped to a matrix
-    """
-    shape = array.shape
-    matrix = array.reshape(shape[0], np.prod(shape[1:]))
-    return matrix
-
-
-# Dictionary mapping from embedding names to Callables
-embedding_dict = {"matrix": array_to_matrix}
-
-
-# Dictionary mapping from kernel names to Callables
-kernel_dict = {
+MMD_KERNEL_DICT = {
     "rbf": metrics.pairwise.rbf_kernel,
     "laplace": metrics.pairwise.laplacian_kernel,
 }
@@ -56,21 +34,21 @@ def mmd(
         float: The maximum mean discrepancy between A and B
     """
     # Extract embedding callable
-    embedding = embedding_dict[embedding_name]
+    embedding_fn = EMBEDDING_FN_DICT[embedding_name]
 
     # In the future, consider optionally using feature embeddings here
-    matrix_A = embedding(array_A)
-    matrix_B = embedding(array_B)
+    matrix_A = embedding_fn(array_A)
+    matrix_B = embedding_fn(array_B)
     N_A = matrix_A.shape[0]
     N_B = matrix_B.shape[0]
 
     # Extract kernel callable
-    kernel = kernel_dict[kernel_name]
+    kernel_fn = MMD_KERNEL_DICT[kernel_name]
 
     # Compute MMD components
-    kernel_AA = kernel(X=matrix_A)
-    kernel_BB = kernel(X=matrix_B)
-    kernel_AB = kernel(X=matrix_A, Y=matrix_B)
+    kernel_AA = kernel_fn(X=matrix_A)
+    kernel_BB = kernel_fn(X=matrix_B)
+    kernel_AB = kernel_fn(X=matrix_A, Y=matrix_B)
 
     # Compute MMD
     mmd = (
